weekly_report/requirements.txt
pylint
pytest
pandas
pytest-cov
python-dotenv
psycopg2-binary
reportlab
xhtml2pdf
boto3